pyproject.toml
[tool.poetry]
name = "gh-stats-gen"
version = "1.0.0"
description = ""
authors = ["BD103 <dont@stalkme>"]
license = "Apache-2.0"

[tool.poetry.dependencies]
python = "^3.8"
PyGithub = "^1.55"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
